refactor: reorganised files so that components have own css files
