+++ src/pages/Signup.js
@@ -1,12 +1,12 @@
 import React from "react";
 
-import SignupForm from "../components/SignupForm";
+import SignupForm from "../components/SignupForm/SignupForm";
 
 const Signup = () => {
   return (
-    <div id="sign-up-page">
-      <div id="sign-up-card">
-        <h2>Welcome to the Chat</h2>
+    <div id="signup-page">
+      <div id="signup-card">
+        <h2>Sign up to the chat</h2>
         <SignupForm />
         <br />
         <div className="text-center my-2">Already have an account?</div>
